+++ app/models/shop.rb
@@ -27,7 +27,7 @@
 class Shop < ApplicationRecord
   paginates_per 30
 
-  searchkick searchable: %i[name shop_type station celestial_object starsystem refinary],
+  searchkick searchable: %i[name shop_type station celestial_object starsystem refinery],
              word_start: %i[name],
              filterable: []
 
@@ -38,7 +38,7 @@ class Shop < ApplicationRecord
       station: station.name,
       celestial_object: station.celestial_object.name,
       starsystem: station.celestial_object.starsystem&.name,
-      refinary: refinary_terminal? ? 'Refinary' : ''
+      refinery: refinery_terminal? ? 'Refinery' : ''
     }
   end
 
@@ -58,7 +58,7 @@ class Shop < ApplicationRecord
   enum shop_type: {
     clothing: 0, armor: 1, weapons: 2, components: 3, armor_and_weapons: 4, superstore: 5,
     ships: 6, admin: 7, bar: 8, hospital: 9, salvage: 10, resources: 11, rental: 12,
-    computers: 13, blackmarket: 14, mining_equipment: 15, equipment: 16, courier: 17, refinary: 18
+    computers: 13, blackmarket: 14, mining_equipment: 15, equipment: 16, courier: 17, refinery: 18
   }
   ransacker :shop_type, formatter: proc { |v| Shop.shop_types[v] } do |parent|
     parent.table[:shop_type]