+++ test/controllers/api/v1/shops_controller_test.rb
@@ -37,7 +37,7 @@ module Api
           'storeImageLarge' => new_deal.store_image.large.url,
           'storeImageMedium' => new_deal.store_image.medium.url,
           'storeImageSmall' => new_deal.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'Port Olisar',
             'slug' => 'port-olisar'
@@ -100,7 +100,7 @@ module Api
           'storeImageLarge' => admin_daymar.store_image.large.url,
           'storeImageMedium' => admin_daymar.store_image.medium.url,
           'storeImageSmall' => admin_daymar.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'ArcCorp 001',
             'slug' => 'arccorp'
@@ -197,7 +197,7 @@ module Api
           'storeImageLarge' => admin_yela.store_image.large.url,
           'storeImageMedium' => admin_yela.store_image.medium.url,
           'storeImageSmall' => admin_yela.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'ArcCorp 002',
             'slug' => 'arccorp'
@@ -294,7 +294,7 @@ module Api
           'storeImageLarge' => admin_olisar.store_image.large.url,
           'storeImageMedium' => admin_olisar.store_image.medium.url,
           'storeImageSmall' => admin_olisar.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'Port Olisar',
             'slug' => 'port-olisar'
@@ -354,7 +354,7 @@ module Api
           'storeImageLarge' => dumpers.store_image.large.url,
           'storeImageMedium' => dumpers.store_image.medium.url,
           'storeImageSmall' => dumpers.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'Port Olisar',
             'slug' => 'port-olisar'
@@ -414,7 +414,7 @@ module Api
           'storeImageLarge' => new_deal.store_image.large.url,
           'storeImageMedium' => new_deal.store_image.medium.url,
           'storeImageSmall' => new_deal.store_image.small.url,
-          'refinaryTerminal' => nil,
+          'refineryTerminal' => nil,
           'station' => {
             'name' => 'Port Olisar',
             'slug' => 'port-olisar'